+++ R/prism_archive_verify.R
@@ -6,9 +6,9 @@
 #' [prism_archive_subset()]. Any files that are not readable can automatically 
 #' be re-downloaded.
 #' 
-#' Under the hood, it uses `raster::stack()` and then `raster::rasterToPoints()`
-#' to determine if the bil files are readable. If both those files are able 
-#' to successfully read the files, they are assumed to be valid/readable.
+#' Under the hood, it uses `terra::rast()` to determine if the PRISM data files 
+#' are readable. If both those files are able to successfully read the files, 
+#' they are assumed to be valid/readable.
 #' 
 #' @inheritParams prism_archive_subset
 #'